misc.c: get32(): use uint32 rval to avoid signed-shift UB
int32 rval <<= 8 is signed overflow once bit 31 is set (any IP >=
128.0.0.0, half of all TCP seqnos, all SNTP timestamps). The old code
was safe because int32 was 'long' on LP64; the LP64-correct int32=int
introduced the regression. UndefinedBehaviorSanitizer flags the last
shift at runtime. Fix: uint32 accumulator, cast to int32 on return.

+++ src/misc.c
@@ -202,7 +202,7 @@ get16(uint8 *cp)
 int32
 get32(uint8 *cp)
 {
-	int32 rval;
+	uint32 rval;
 
 	rval = *cp++;
 	rval <<= 8;
@@ -212,7 +212,7 @@ get32(uint8 *cp)
 	rval <<= 8;
 	rval |= *cp;
 
-	return rval;
+	return (int32) rval;
 }
 /* Compute int(log2(x)) */
 int
